requirements.txt
amqp                          5.3.1
asgiref                       3.8.1
async-timeout                 5.0.1
bcrypt                        4.3.0
billiard                      4.2.1
celery                        5.4.0
click                         8.1.8
click-didyoumean              0.3.1
click-plugins                 1.1.1
click-repl                    0.3.0
colorama                      0.4.6
cron-descriptor               1.4.5
dataclasses                   0.6
Django                        5.1.7
django-celery-beat            2.7.0
django-redis                  5.4.0
django-timezone-field         7.1
djangorestframework           3.15.2
djangorestframework_simplejwt 5.5.0
djongo                        1.2.31
kombu                         5.5.1
MarkupSafe                    3.0.2
mongoengine                   0.29.1
pip                           25.0.1
prompt_toolkit                3.0.50
PyJWT                         2.9.0
pymongo                       3.11.4
python-crontab                3.2.0
python-dateutil               2.9.0.post0
pytz                          2025.2
redis                         5.2.1
setuptools                    75.8.2
six                           1.17.0
sqlparse                      0.5.3
typing_extensions             4.13.0
tzdata                        2025.1
vine                          5.1.0
wcwidth                       0.2.13
Werkzeug                      3.1.3
wheel                         0.45.1
